Vss2Git

+++ Source/OldCode/Archive/2008 TVA Code Library/Source/PCS.System/PrecisionTimer.cs
@@ -175,7 +175,7 @@ namespace System
                 {
                     long elapsedTicks = m_stopwatch.ElapsedTicks;
                     ImmutableTimeState timeState = m_timeState;
-                    DateTime precisionTime = timeState.BaseTime.AddTicks(((int)((elapsedTicks - timeState.ElapsedTicks) * Ticks.PerSecond) / timeState.SystemFrequency));
+                    DateTime precisionTime = timeState.BaseTime.AddTicks(((elapsedTicks - timeState.ElapsedTicks) * Ticks.PerSecond) / timeState.SystemFrequency);
 
                     if (elapsedTicks >= timeState.ElapsedTicks + m_synchronizePeriodStopwatchTicks)
                     {
@@ -183,8 +183,8 @@ namespace System
                         DateTime systemTime = DateTime.UtcNow;
 
                         // Last parameter is a calculation that asymptotically approachs the measured system frequency
-                        m_timeState = new ImmutableTimeState(systemTime, precisionTime, elapsedTicks, (int)(((elapsedTicks - timeState.ElapsedTicks) * Ticks.PerSecond * 2) /
-                            (systemTime.Ticks - timeState.ObservedTime.Ticks + systemTime.Ticks + systemTime.Ticks - precisionTime.Ticks - timeState.ObservedTime.Ticks)));
+                        m_timeState = new ImmutableTimeState(systemTime, precisionTime, elapsedTicks, ((elapsedTicks - timeState.ElapsedTicks) * Ticks.PerSecond * 2) /
+                            (systemTime.Ticks - timeState.ObservedTime.Ticks + systemTime.Ticks + systemTime.Ticks - precisionTime.Ticks - timeState.ObservedTime.Ticks));
                     }
 
                     // Return high-resolution timestamp
